Add files via upload
My name is Harshita, and I’m a UI/UX designer.
For this project, my goal was to redesign the YouTube comment section to dynamically update based on what the viewer is currently watching in the video. The focus was on creating a clean, easy-to-read interface especially optimized for smaller screens like mobile devices.

To enhance clarity and interaction, I introduced a pin icon system on the video’s progress bar. These pins indicate where comments appear throughout the video. By default, the pins are white, while the currently active comment , the one being engaged with is highlighted in red for easy visibility.

To support intuitive navigation, each comment is accompanied by a timestamp and a video preview thumbnail, helping users quickly understand the context of the comment without needing to scrub through the entire video.

Finally, to improve readability and reduce clutter, replies are organized in a threaded format. This structure makes discussions easier to follow and creates a more visually appealing and seamless user experience.